chart

+++ Backend/Tar/soundpadtest.py
@@ -6,7 +6,7 @@ import time
 from PyQt5.QtWidgets import (
     QApplication, QWidget, QPushButton, QDesktopWidget, QFileDialog,
     QTableWidget, QTableWidgetItem, QLabel, QMainWindow, QFormLayout,
-    QGroupBox, QScrollArea, QVBoxLayout, QHBoxLayout, QProgressDialog, QLineEdit, QShortcut
+    QGroupBox, QScrollArea, QVBoxLayout, QHBoxLayout, QProgressDialog, QLineEdit, QMessageBox
 )
 from PyQt5.QtGui import QIcon, QPixmap
 from PyQt5.QtMultimedia import QMediaPlayer, QMediaContent
@@ -108,16 +108,17 @@ class App(QWidget):
         folder = r""
       
         # เห็นเฉพาะ .wav, .mp3
-        fname, _ = QFileDialog.getOpenFileName(self, "QFileDialog.getOpenFileName()", folder, "WAV Files (*.wav);; MP3 Files (*.mp3)", options=options)
-    
-        # new_fname = self.check_filename(fname)
-        # if new_fname != fname:
-        #     print("file name exists, renamed to", new_fname)
-        #     fname = new_fname
+        fname, _ = QFileDialog.getOpenFileName(self, "QFileDialog.getOpenFileName()", folder, "WAV Files (*.wav);; MP3 Files (*.mp3)", options=options) 
 
         if fname:
-            print("add file :", fname)
+            if self.check_duplicate_file(fname):
+                msg_box = QMessageBox()
+                msg_box.setText("File name already exists.")
+                msg_box.exec_()
+                return
             
+            print("add file :", fname)
+
             row = self.table.rowCount()
             self.table.insertRow(row)
             self.table.setItem(row, 0, QTableWidgetItem(os.path.basename(fname)))
@@ -140,19 +141,11 @@ class App(QWidget):
             self.filenames.append(fname)
             self.save_file()
 
-            # if not all(os.path.exists(f) for f in self.filenames):
-            #     self.filenames = [f for f in self.filenames if os.path.exists(f)]
-            #     self.save_file()
-
-    def check_filename(self, filename):
-        new_filename = filename
-        count = 1
-        while os.path.exists(new_filename):
-            basename, ext = os.path.splitext(filename)
-            new_basename = f"{basename}({count})"
-            new_filename = f"{new_basename}{ext}"
-            count += 1
-        return new_filename
+    def check_duplicate_file(self, file_path):
+        file_name = os.path.basename(file_path)
+        if file_name in set([os.path.basename(fname) for fname in self.filenames]):
+            return True
+        return False
 
     def save_file(self):
         data = {}
@@ -165,12 +158,6 @@ class App(QWidget):
         
         print("save success")
         print(data)
-
-        # data = dict(sorted(data.items(), key=lambda x: x[1], reverse=True))
-        # # data[fname.split("/")[-1].split(".")[0]] = self.play_counts[fname]
-        # with open('sort_counts.txt', 'w') as f:
-        #     for fname, count in data.items():
-        #         f.write(f"{fname}: {count}\n")
 
         sort_counts = sorted(self.play_counts.items(), key=lambda x: x[1], reverse=True)
         with open("sort_counts.txt", "w") as file:
